backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-socketio>=5.16.0",
    "uvicorn[standard]>=0.40.0",
]
